requirements.txt
aiohttp==3.6.2
arrow==0.17.0
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
Babel==2.9.0
certifi==2020.4.5.1
cffi==1.14.3
chardet==3.0.4
choice==0.1
clipboard==0.0.4
DateTime==4.3
debtcollector==2.2.0
discord.py==1.8.2
dnspython==1.16.0
eventlet==0.29.1
fasteners==0.15
fuzzywuzzy==0.18.0
geojson==2.5.0
greenlet==0.4.17
idna==2.9
logger==1.4
loguru==0.5.3
importlib-metadata==2.0.0
iso8601==0.1.13
monotonic==1.5
mouse==0.7.1
msgpack==1.0.0
multidict==4.7.5
netaddr==0.8.0
netifaces==0.10.9
numpy==1.19.4
os-win==5.2.0
oslo.concurrency==4.3.1
oslo.config==8.3.2
oslo.context==3.1.1
oslo.i18n==5.0.1
oslo.log==4.4.0
oslo.serialization==4.0.1
oslo.utils==4.7.0
packaging==20.4
pafy==0.5.5
pause==0.3
pbr==5.5.1
pyparsing==2.4.7
pyperclip==1.8.1
PyQt5==5.14.2
PyQt5-sip==12.7.2
PySide2==5.14.2.1
PySocks==1.7.1
pyTelegramBotAPI==3.7.1
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.25.1
rfc3986==1.4.0
setuptools==41.2.0
shiboken2==5.14.2.1
six==1.14.0
SpeechRecognition==3.8.1
stevedore==3.2.2
times==0.7
urllib3==1.25.9
urwid==2.1.2
websockets==8.1
wrapt==1.12.1
yarl==1.4.2
youtube-dl==2020.11.12
zipp==3.4.0
zope.interface==5.2.0